requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.4
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.20
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2